+++ src/lib/components/delegations/DelegationInfo.tsx
@@ -1,6 +1,7 @@
 import { Button, Flex, Heading } from "@chakra-ui/react";
 import type { MouseEventHandler } from "react";
 
+import { TableTitle } from "../table";
 import { AmpEvent, track } from "lib/amplitude";
 import { useMobile } from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
@@ -21,8 +22,8 @@ export const DelegationInfo = ({
   onViewMore,
 }: DelegationInfoProps) => {
   const isMobile = useMobile();
-
   const isMobileOverview = isMobile && !!onViewMore;
+
   return (
     <>
       {isMobileOverview ? (
@@ -43,16 +44,8 @@ export const DelegationInfo = ({
           <CustomIcon name="chevron-right" color="gray.600" />
         </Flex>
       ) : (
-        <Flex direction="column" gap={4}>
-          <Heading
-            variant="h6"
-            as="h6"
-            display="flex"
-            alignItems="center"
-            height={{ base: "29px", md: "auto" }}
-          >
-            Delegations
-          </Heading>
+        <Flex direction="column" gap={4} mt={{ base: 4, md: 0 }}>
+          <TableTitle title="Delegations" mb={0} showCount={false} />
           <Flex
             direction={{ base: "column", md: "row" }}
             alignItems={{ base: "start", md: "center" }}