+++ src/lib/components/select-code/CodeSelectDrawerButton.tsx
@@ -69,8 +69,6 @@ export const CodeSelectDrawerButton = ({
     onClose();
   };
 
-  const isSearching = !!keyword || permissionValue !== "all";
-
   return (
     <>
       <Button
@@ -119,26 +117,26 @@ export const CodeSelectDrawerButton = ({
             <Tabs px={6}>
               <TabList borderBottom="1px" borderColor="gray.800">
                 <CustomTab count={storedCodesCount}>My Stored Codes</CustomTab>
-                <CustomTab count={savedCodesCount}>My Saved Codes </CustomTab>
+                <CustomTab count={savedCodesCount}>My Saved Codes</CustomTab>
               </TabList>
               <TabPanels>
                 <TabPanel p={0}>
                   <MyStoredCodesTable
                     codes={stored}
+                    totalData={storedCodesCount}
                     isLoading={isStoredCodesLoading}
                     onRowSelect={handleSelect}
                     emptyMessage="You don’t have any stored codes in this device."
                     disconnectedMessage="to see your stored code."
-                    isSearching={isSearching}
                     isReadOnly
                   />
                 </TabPanel>
                 <TabPanel p={0}>
                   <MySavedCodesTable
                     codes={saved}
+                    totalData={savedCodesCount}
                     isLoading={isSavedCodesLoading}
                     onRowSelect={handleSelect}
-                    isSearching={isSearching}
                     isReadOnly
                   />
                 </TabPanel>