+++ src/lib/pages/account-details/types.ts
@@ -1,7 +1,6 @@
 import { z } from "zod";
 
 import { zAddr } from "lib/types";
-import type { TokenWithValue, Validator } from "lib/types";
 
 export enum TabIndex {
   Overview = "overview",
@@ -27,32 +26,3 @@ export const zAccountDetailsQueryParams = z.object({
       .transform(() => TabIndex.Overview),
   ]),
 });
-
-export interface StakingParams {
-  unbondingTime: string; // e.g. "14 days"
-  maxEntries: number;
-  bondDenoms: TokenWithValue[];
-}
-
-export interface Delegation {
-  validator: Validator;
-  balances: TokenWithValue[];
-}
-
-export interface Unbonding {
-  validator: Validator;
-  completionTime: Date;
-  balances: TokenWithValue[];
-}
-
-export interface Redelegation {
-  srcValidator: Validator;
-  dstValidator: Validator;
-  completionTime: Date;
-  balances: TokenWithValue[];
-}
-
-export interface NonRedelegatable {
-  dstValidator: Validator;
-  completionTime: Date;
-}