+++ src/lib/pages/tx-details/components/tx-message/index.tsx
@@ -3,14 +3,13 @@ import type { logs } from "@cosmjs/stargate";
 import { useState } from "react";
 
 import type { MsgBody } from "lib/services/tx";
-import type { Option } from "lib/types";
 
 import { TxMsgDetails } from "./TxMsgDetails";
 import { TxMsgExpand } from "./TxMsgExpand";
 
 export interface TxMsgData {
   msgBody: MsgBody;
-  log: Option<logs.Log>;
+  log: logs.Log;
   isSingleMsg?: boolean;
 }