+++ CHANGELOG.md
@@ -40,14 +40,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - [#727](https://github.com/alleslabs/celatone-frontend/pull/727) Add amp nft pages
+- [#720](https://github.com/alleslabs/celatone-frontend/pull/720) Add delegations for contract detail
 - [#684](https://github.com/alleslabs/celatone-frontend/pull/684) Add nft page
+- [#717](https://github.com/alleslabs/celatone-frontend/pull/717) Add total value for contract detail
 - [#711](https://github.com/alleslabs/celatone-frontend/pull/711) Refactor assetInfos and add movePoolInfos to tx details
+- [#724](https://github.com/alleslabs/celatone-frontend/pull/724) Add stone-13
 
 ### Improvements
 
+- [#728](https://github.com/alleslabs/celatone-frontend/pull/728) Support tx events for cosmos sdk 0.50
+- [#729](https://github.com/alleslabs/celatone-frontend/pull/729) Add missing title for block transactions
+- [#726](https://github.com/alleslabs/celatone-frontend/pull/726) Add missing title for instantiate permission and adjust contract menu in nav bar
+- [#713](https://github.com/alleslabs/celatone-frontend/pull/713) Adjust search state consistency
+- [#712](https://github.com/alleslabs/celatone-frontend/pull/712) api v1 - proposal list
+- [#714](https://github.com/alleslabs/celatone-frontend/pull/714) api v1 - recent contracts list
 - [#710](https://github.com/alleslabs/celatone-frontend/pull/710) Refactor all address types
 
 ### Bug fixes
+
+- [#722](https://github.com/alleslabs/celatone-frontend/pull/722) Fix flicker query page
 
 ## v1.4.1